+++ repo/model/response.go
@@ -1,6 +1,6 @@
 package model
 
 type Response struct {
-	Code int    `form:"code" json:"code"`
+	Code int64  `form:"code" json:"code"`
 	Msg  string `form:"msg" json:"msg"`
 }
